Add files via upload
"This Guided Project "Python for Beginners: Variables and Strings" is for individuals who are new to programming and want to learn the basics of Python. In this 2-hour long project-based course, you will learn how to work with Python strings, perform string manipulations using built-in methods and apply these skills to real-world scenarios through a shopping list cost analysis short project. This guided-project will cover the learning objectives of working with Python strings, performing string manipulation using built-in methods and applying these skills in real-world scenarios. This project is designed to provide a comprehensive introduction to variables and strings in a beginner-friendly format. In order to be successful in this project, you will need to have prior knowledge of basic Python concepts; Familiarity with control structures in Python such as for loops and if statements."